# server/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Server for Legacy App",
        "title": "Legacy API",
        "contact": {
            "name": "David Oduneye and Akshay Dupuguntla",
            "email": "oduneye.d@northeastern.edu and dupuguntla@northeastern.edu"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/progresses/subtask/{uid}/{sid}": {
            "get": {
                "description": "Get progress of a specific subtask for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get subtask progress",
                "operationId": "get-subtask-progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Subtask ID",
                        "name": "sid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SubTaskProgress"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch subtask progress",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/progresses/subtask/{uid}/{sid}/complete": {
            "put": {
                "description": "Mark a subtask as completed for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Complete subtask progress",
                "operationId": "complete-subtask-progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Subtask ID",
                        "name": "sid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SubTaskProgress"
                        }
                    },
                    "404": {
                        "description": "Failed to complete subtask progress",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/progresses/task/{uid}/{tid}": {
            "get": {
                "description": "Get progress of a specific task for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get task progress",
                "operationId": "get-task-progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TaskProgress"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch task progress",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/progresses/{uid}/{tid}": {
            "get": {
                "description": "Get all subtask progresses for a specific task of a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress"
                ],
                "summary": "Get all subtask progresses of a task",
                "operationId": "get-all-subtask-progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SubTaskProgress"
                            }
                        }
                    },
                    "404": {
                        "description": "Failed to fetch all subtask progresses of task",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/subtasks/{sid}": {
            "get": {
                "description": "Returns a subtask from subtask id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subtask"
                ],
                "summary": "Gets a subtask from subtask id",
                "operationId": "get-subtask",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SubTaskID",
                        "name": "sid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SubTask"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch subtask",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a subtask",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subtask"
                ],
                "summary": "Deletes a subtask",
                "operationId": "delete-subtask",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SubTaskID",
                        "name": "sid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subtask successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Failed to delete subtask",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a subtask",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subtask"
                ],
                "summary": "Updates a subtask",
                "operationId": "update-subtask",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SubTaskID",
                        "name": "sid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SubTask",
                        "name": "subtask",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.SubTaskDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SubTask"
                        }
                    },
                    "400": {
                        "description": "Failed to update subtask",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/tasks/{tid}": {
            "get": {
                "description": "Returns a task from task id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Gets a task from task id",
                "operationId": "get-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TaskID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch task",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Deletes a task",
                "operationId": "delete-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TaskID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Failed to delete task",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Updates a task",
                "operationId": "update-task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TaskID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.TaskDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    },
                    "400": {
                        "description": "Failed to update task",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docmodels.SubTaskDTO": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "actions_state": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subtask_description": {
                    "type": "string"
                },
                "subtask_name": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                }
            }
        },
        "docmodels.TaskDTO": {
            "type": "object",
            "properties": {
                "personas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "task_description": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                }
            }
        },
...
        "models.File": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "object_key": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Guide": {
            "type": "object",
            "required": [
                "author",
                "author_image_url",
                "full_text",
                "guide_name",
                "sub_title",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "author_image_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "date": {
                    "type": "string"
                },
                "full_text": {
                    "type": "string"
                },
                "guide_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "mins_read": {
                    "type": "integer"
                },
                "sub_title": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.Persona": {
            "type": "object",
            "required": [
                "persona_description",
                "persona_title"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "persona_description": {
                    "type": "string"
                },
                "persona_title": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Task"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
...
        "models.SubTask": {
            "type": "object",
            "required": [
                "sub_task_description",
                "sub_task_name",
                "task_id"
            ],
            "properties": {
                "actions": {
                    "type": "string"
                },
                "actions_state": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "sub_task_description": {
                    "type": "string"
                },
                "sub_task_name": {
                    "type": "string"
                },
                "task_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.SubTaskProgress": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "sub_task_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Tag": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.File"
                    }
                },
                "guides": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Guide"
                    }
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Task"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.Task": {
            "type": "object",
            "required": [
                "task_description",
                "task_name"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "personas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Persona"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "task_description": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.TaskProgress": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "progress": {
                    "type": "integer"
                },
                "task_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}